backend/pyproject.toml
[project]
name = "backend"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "flask>=3.1.0",
    "flask-cors>=5.0.1",
    "google-cloud-vision>=3.10.1",
    "pillow>=11.1.0",
    "python-dotenv>=1.1.0",
    "requests>=2.32.3",
    "transformers>=4.50.3",
]
